Prepare folder structure for gleaner
