+++ .gitignore
@@ -1,7 +1,6 @@
 **/secret/
 secret/**
 logs/**
-nabu
 **/*.log
 **/.idea/**
 **/.fleet/**
